refactor: split adminshop command
