+++ slash_commands/subcommands/adminshop/index.js
@@ -1,0 +1,9 @@
+const { cancel } = require("./cancel");
+const { deleteItem } = require("./delete");
+const { refund } = require("./refund");
+
+module.exports = {
+    cancel,
+    deleteItem,
+    refund,
+};
